requirements.txt
pyaudio
websockets
openai
requests
pydub
python-dotenv
speechrecognition
ffmpeg-python
pyttsx3
transformers
datasets
pinecone-client
google-generativeai
nltk


